Use ConversationMediaView to simplify media rendering in conversation view cells.
